+++ hub-saml/src/main/java/uk/gov/ida/saml/hub/validators/response/common/ResponseDestinationValidator.java
@@ -1,4 +1,4 @@
-package uk.gov.ida.saml.hub.transformers.inbound.decorators;
+package uk.gov.ida.saml.hub.validators.response.common;
 
 import org.opensaml.saml.saml2.core.Response;
 import uk.gov.ida.saml.core.validators.DestinationValidator;
