+++ hub-saml/src/main/java/uk/gov/ida/saml/hub/validators/response/common/ResponseMaxSizeValidator.java
@@ -1,4 +1,4 @@
-package uk.gov.ida.saml.hub.transformers.inbound.decorators;
+package uk.gov.ida.saml.hub.validators.response.common;
 
 import javax.inject.Inject;
 import uk.gov.ida.saml.hub.validators.StringSizeValidator;
